# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "EBYS - İş Akışı ve Onay Yönetim Sistemi API",
    "description" : "Uçların çoğu kimlik doğrulaması ister.\n\nKullanım: `/api/auth/login` ucundan e-posta ve şifreyle giriş yapın,\ndönen `accessToken` değerini sağ üstteki **Authorize** butonuna yapıştırın.\nBundan sonraki isteklere `Authorization: Bearer <token>` başlığı otomatik eklenir.\n",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/api/records/{id}" : {
      "get" : {
        "tags" : [ "record-controller" ],
        "operationId" : "getRecordById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecordResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "record-controller" ],
        "operationId" : "updateRecord",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecordUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecordResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "record-controller" ],
        "operationId" : "deleteRecord",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/records" : {
      "get" : {
        "tags" : [ "record-controller" ],
        "operationId" : "getAllRecords",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "TASLAK", "BSK_YRD_INCELEMESINDE", "BASKAN_INCELEMESINDE", "DUZENLEME_BEKLIYOR", "ONAYLANDI", "REDDEDILDI" ]
          }
        }, {
          "name" : "categoryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "creator",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseRecordSearchResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "record-controller" ],
        "operationId" : "createRecord",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecordCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecordResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audit-logs/record/{recordId}" : {
      "get" : {
        "tags" : [ "audit-log-controller" ],
        "operationId" : "getGecmis_1",
        "parameters" : [ {
          "name" : "recordId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AuditLogResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "RecordUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "minLength" : 1
          },
          "description" : {
            "type" : "string",
            "minLength" : 1
          },
          "categoryId" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "categoryId", "description", "title" ]
      },
      "RecordResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "categoryId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TASLAK", "BSK_YRD_INCELEMESINDE", "BASKAN_INCELEMESINDE", "DUZENLEME_BEKLIYOR", "ONAYLANDI", "REDDEDILDI" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdByFullName" : {
            "type" : "string"
          },
          "assignment" : {
            "$ref" : "#/components/schemas/AssignmentView"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RecordCreateRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "minLength" : 1
          },
          "description" : {
            "type" : "string",
            "minLength" : 1
          },
          "categoryId" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "categoryId", "description", "title" ]
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "PagedResponseRecordSearchResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RecordSearchResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RecordSearchResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "categoryId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TASLAK", "BSK_YRD_INCELEMESINDE", "BASKAN_INCELEMESINDE", "DUZENLEME_BEKLIYOR", "ONAYLANDI", "REDDEDILDI" ]
          },
          "createdBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdByFullName" : {
            "type" : "string"
          },
          "assignedTo" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "assignment" : {
            "$ref" : "#/components/schemas/AssignmentView"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "AuditLogResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "recordId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userFullName" : {
            "type" : "string"
          },
          "roleId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "roleName" : {
            "type" : "string"
          },
          "action" : {
            "type" : "string"
          },
          "previousStatus" : {
            "type" : "string"
          },
          "newStatus" : {
            "type" : "string"
          },
          "comment" : {
            "type" : "string"
          },
          "previousAssignment" : {
            "$ref" : "#/components/schemas/AssignmentView"
          },
          "newAssignment" : {
            "$ref" : "#/components/schemas/AssignmentView"
          },
          "httpMethod" : {
            "type" : "string"
          },
          "requestPath" : {
            "type" : "string"
          },
          "httpStatus" : {
            "type" : "integer",
            "format" : "int32"
          },
          "errorCode" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "AssignmentView" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "type" : "string",
            "enum" : [ "USER", "DEPARTMENT", "NONE" ]
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userFullName" : {
            "type" : "string"
          },
          "departmentId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "departmentName" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Giriş sonrası dönen accessToken",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}